requirements.txt
absl-py==2.1.0
accelerate==1.5.2
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.6.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
ascendctools==0.1.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrdict==2.0.1
attrs==25.1.0
auto-tune==0.1.0
babel==2.17.0
beautifulsoup4==4.13.3
binpacking==1.5.2
black==25.1.0
bleach==6.2.0
boto3==1.37.13
botocore==1.37.13
cachier==3.1.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cos-python-sdk-v5==1.9.34
cpm-kernels==1.0.11
crcmod==1.7
cryptography==43.0.3
cycler==0.12.1
dacite==1.9.2
dataflow==0.0.1
datasets==3.1.0
debugpy==1.8.12
decorator==5.2.0
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.8
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.17.0
flatbuffers==25.2.10
flwr==1.15.2
flwr-datasets==0.5.0
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
fvcore==0.1.5.post20221221
gradio==5.21.0
gradio_client==1.7.2
groovy==0.1.2
grpcio==1.70.0
h11==0.14.0
hccl==0.1.0
hccl-parser==0.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
iopath==0.1.10
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
isort==6.0.1
iterators==0.0.2
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.5
jiter==0.9.0
jmespath==0.10.0
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lightgbm==4.6.0
lightning==2.5.0.post0
lightning-utilities==0.12.0
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
modelscope==1.23.2
mpmath==1.3.0
ms_swift==3.2.1
msadvisor==1.0.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
nvidia-ml-py==12.570.86
onnx==1.17.0
onnxruntime==1.21.0
op-gen==0.1
op-test-frame==0.1
opc-tool==0.1.0
openai==1.66.3
opencv-python==4.11.0.86
opt-einsum==3.3.0
orjson==3.10.15
oss2==2.19.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
peewee==3.17.9
peft==0.14.0
pexpect==4.9.0
pillow==11.1.0
pip==25.0.1
pipx==1.7.1
platformdirs==4.3.6
portalocker==3.1.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.20
pytorch-lightning==2.5.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
ray==2.31.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rouge==1.0.1
rpds-py==0.23.1
ruff==0.11.2
s3transfer==0.11.4
safehttpx==0.1.6
safetensors==0.5.2
schedule-search==0.0.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
setuptools==69.5.1
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
starlette==0.45.3
swanboard==0.1.7b1
swankit==0.1.6
swanlab==0.5.3
sympy==1.13.3
tabulate==0.9.0
te==0.4.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.9.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.20.3
tomli==2.2.1
tomli_w==1.2.0
tomlkit==0.13.2
torch==2.2.0
torch-npu==2.2.0
torchaudio==2.2.0
torchmetrics==1.6.2
torchvision==0.17.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.0
transformers-stream-generator==0.0.5
trl==0.15.2
typer==0.12.5
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
uri-template==1.3.0
urllib3==2.3.0
userpath==1.9.2
uv==0.6.9
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
xmltodict==0.14.2
xxhash==3.5.0
yacs==0.1.8
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
